requirements.txt
streamlit==1.28.0
pandas==2.1.3
neo4j==5.14.0
pymilvus==2.3.4
pypdf2==3.0.1
pdfplumber==0.11.0
sentence-transformers==2.3.1
openai==1.3.7
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1

